+++ lib/modules/posts/schemas/post_like.ex
@@ -24,7 +24,7 @@ defmodule PhoenixKit.Modules.Posts.PostLike do
 
   @type t :: %__MODULE__{
           uuid: UUIDv7.t() | nil,
-          post_id: UUIDv7.t(),
+          post_uuid: UUIDv7.t(),
           user_id: integer() | nil,
           user_uuid: UUIDv7.t() | nil,
           post: PhoenixKit.Modules.Posts.Post.t() | Ecto.Association.NotLoaded.t(),
@@ -34,7 +34,10 @@ defmodule PhoenixKit.Modules.Posts.PostLike do
         }
 
   schema "phoenix_kit_post_likes" do
-    belongs_to :post, PhoenixKit.Modules.Posts.Post, references: :uuid, type: UUIDv7
+    belongs_to :post, PhoenixKit.Modules.Posts.Post,
+      foreign_key: :post_uuid,
+      references: :uuid,
+      type: UUIDv7
 
     belongs_to :user, PhoenixKit.Users.Auth.User,
       foreign_key: :user_uuid,
@@ -60,11 +63,11 @@ defmodule PhoenixKit.Modules.Posts.PostLike do
   """
   def changeset(like, attrs) do
     like
-    |> cast(attrs, [:post_id, :user_id, :user_uuid])
-    |> validate_required([:post_id, :user_uuid])
-    |> foreign_key_constraint(:post_id)
+    |> cast(attrs, [:post_uuid, :user_id, :user_uuid])
+    |> validate_required([:post_uuid, :user_uuid])
+    |> foreign_key_constraint(:post_uuid)
     |> foreign_key_constraint(:user_uuid)
-    |> unique_constraint([:post_id, :user_id],
+    |> unique_constraint([:post_uuid, :user_id],
       name: :phoenix_kit_post_likes_post_id_user_id_index,
       message: "you have already liked this post"
     )