+++ lib/modules/tickets/ticket.ex
@@ -126,11 +126,11 @@ defmodule PhoenixKit.Modules.Tickets.Ticket do
     field :user_id, :integer
     field :assigned_to_id, :integer
 
-    has_many :comments, PhoenixKit.Modules.Tickets.TicketComment, foreign_key: :ticket_id
-    has_many :attachments, PhoenixKit.Modules.Tickets.TicketAttachment, foreign_key: :ticket_id
+    has_many :comments, PhoenixKit.Modules.Tickets.TicketComment, foreign_key: :ticket_uuid
+    has_many :attachments, PhoenixKit.Modules.Tickets.TicketAttachment, foreign_key: :ticket_uuid
 
     has_many :status_history, PhoenixKit.Modules.Tickets.TicketStatusHistory,
-      foreign_key: :ticket_id
+      foreign_key: :ticket_uuid
 
     timestamps(type: :utc_datetime)
   end